DRILL-1170 - 5. AM Shell
DRILL-1170 - Drill-on-YARN. See JIRA entry for details.
Contains the “shell” of the DoY Application Master (AM). The AM is the
server that YARN launches. The AM manages the Drill cluster.
The code here corresponds to the “App Master Overview” class diagram in
the internal documentation.

PR 5 of 10. (Revised after reformatting)