DRILL-1170 - 9. Tests
DRILL-1170 - Drill-on-YARN. See JIRA entry for details.
Unit tests. Testing DoY with unit tests is challenging: YARN itself is
a cluster manager, which is hard to run in-process. The unit tests
cover various items that can be tested in process. One test
exhaustively covers the recent script changes.

PR 9 of 10.