MD-1144: Drill-on-YARN timeout issues
Fixes a concurrent modification issue when a request times out.
Fixes a failure to release a container when a request times out.
Fixes a log message typo.
Added more logging during Drillbit registration.
Fixes an error when polling the ZK registry to see if a Drillbit is
still registered. Also adds logging and Javadoc.
Fix to serialize container requests