DRILL-4181: Improve memory limit assignment for Sort operators
A key challenge for memory limits is to know how much memory to give to
each query. This, in turn, depends on the number of queries to run. 1?
10? 100? If we run 20 queries, say, then, ceteris paribus, each query
should get 1/20th of memory.

We need a way to learn the number of active queries. In particular, we
need a way to LIMIT the number since we can’t change the memory
allocation once a planning decision is made, if it turns out that 100
more queries arrive.

The answer is to use the existing Zookeeper queues to set a limit on
the number of active queries. We already know the total direct memory
available. A bit of math tells us how much memory to assign per
admitted query.

The next step is to adapt the existing memory planning code to use the
memory limit as the amount of memory to give each query.

At the implementation level, the code for queues and memory assignment
is factored out into a new resource manager. The code provides three
versions: null (no query, old-style memory assignment), existing
Zookeeper, and a new single-node, in-process queue for testing. An RM
framework lets advanced users create their own queueing solution.

All of this entailed moving existing code around to set things right
for the above work.