DRILL-5496, MD-1671, MD-1631: Fix for failed Hive connection
If the Hive server restarts, Drill either hangs or continually reports
errors when retrieving schemas. The problem is that the Hive plugin tries
to handle connection failures, but does not do so correctly for the secure
connection case. The problem is complex, see DRILL-5496 for details.

This is a workaround: we discard the entire Hive schema cache when we
encounter an unhandled connection exception, then we rebuild a new one.
This is not a proper fix; for that we'd have to restructure the code.
This will, however, solve the immediate customer case.