DRILL-6953: Convert JSON reader to row set framework
Introduces a new implementation of the JSON reader used when reading
JSON files. Retains the original JSON reader for use in some unit test
and when processing a SELECT ... FROM VALUES ... query.

The revised JSON reader uses an explicit state machine with a separate
state object for each column, and a separate state class for each column
type. This is far easier to understand and debug than the "states via
flags" approach of the original reader.

Uses column information and projection hints to choose the column type.
(Projection hints are used when a column is all nulls.) Defers selecting
a column type as long as possible when the first rows contain only
nulls.

JSON benefits from knowning that a column is not projected before trying
to create the column in the result set loader. Exposed projection
information to allow JSON to create a dummy parse state that
"free-wheels" over all data for that column, avoiding unnecessary schema
conflicts in the reader for unprojected columns.

Adds unit tests and reworks some existing tests.