DRILL-7458: Base storage plugin and filter push-down framework
Provides a "Base" framework for storage plugins similar to the
existing "Easy" framework for format plugins.

Provides a standard filter push-down framework that handles
common work, allowing the storage plugin to implement a
"listener" that applies the push-down.

Provides a "Dummy" storage plugin that acts as an example
and as a "test mule" for the above classes.